requirements.txt
Flask
pandas
python-docx
langchain
langchain_community
langchain-chroma
langchain-huggingface
langchain-groq